+++ javaScript/projects/knights-travails/src/index.js
@@ -1,5 +1,8 @@
+import knightMoves from './containers/knight_travails';
 import test from './components/test';
 import './index.css';
+
+window.knightMoves = knightMoves;
 
 test.print();
 document.body.appendChild(test.render());