+++ javaScript/projects/knights-travails/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 ---
+### 30 JAN 2024
+- 
+---
+### 29 JAN 2024
+- Added more brainstorming comments.
+- The `knightMoves` function will now record all possible first moves.
+- Created `checkArguments` function that will validate inputs for the `knightMoves` function.
+- Created `moveVertically` and `moveHorizontally` functions to move a knight within the game board size.
+---
 ### 14 JAN 2024
 - Reformatted `CHANGELOG.md`.
 - Updated instructions in `README.md`.